Update .gitignore
It's a plain text file that lists patterns of files or directories that Git should exclude from version control. This is useful for excluding files that are generated as part of the build process, configuration files with sensitive information, log files, and other files that don't need to be stored in the repository.